Reorganize the structure of source code
