+++ inference.py
@@ -1,5 +1,5 @@
-from utils import parse_argument
 from accelerate.utils import set_seed
+from llm_box.utils import parse_argument
 
 
 def main():
@@ -10,7 +10,7 @@ def main():
     set_seed(args.seed)
     # TODO: init logger
 
-    from evaluator import Evaluator
+    from llm_box.evaluator import Evaluator
     evaluator = Evaluator(args)
     evaluator.evaluate()
